Move provider utils to util
