+++ packages/block/package.json
@@ -43,7 +43,6 @@
     "@ethereumjs/trie": "^5.0.4",
     "@ethereumjs/tx": "^4.1.1",
     "@ethereumjs/util": "^8.0.5",
-    "cross-fetch": "^3.1.5",
     "ethereum-cryptography": "^1.1.2"
   },
   "devDependencies": {
